refactor(about): extract presentation page into its own comp
