+++ src/app/about/_about-theme.scss
@@ -1,19 +1,14 @@
-@use 'animations';
-@use 'profile-picture/profile-picture-theme';
-@use 'contact-social-icons/contact-social-icons-theme';
-@use 'contact-traditional-icons/contact-traditional-icons-theme';
+@use 'presentation/presentation-theme';
 
 @mixin color($theme) {
   app-about {
-    @include profile-picture-theme.color($theme);
-    @include contact-traditional-icons-theme.color($theme);
-    @include contact-social-icons-theme.color($theme);
+    @include presentation-theme.color($theme);
   }
 }
 
 @mixin motion() {
   app-about {
-    @include profile-picture-theme.motion();
+    @include presentation-theme.motion();
   }
 }
