+++ src/app/about/about.component.ts
@@ -1,20 +1,8 @@
-import { Component, Inject } from '@angular/core'
-import { METADATA } from '../common/injection-tokens'
-import { DescriptionLine, Metadata } from '../metadata'
+import { Component } from '@angular/core'
 
 @Component({
   selector: 'app-about',
   templateUrl: './about.component.html',
   styleUrls: ['./about.component.scss'],
 })
-export class AboutComponent {
-  public readonly realName = this.metadata.realName
-  public readonly nickname = this.metadata.nickname
-  public readonly title = this.metadata.title
-  public readonly rootLine = new DescriptionLine(
-    undefined,
-    this.metadata.descriptionLines,
-  )
-
-  constructor(@Inject(METADATA) private metadata: Metadata) {}
-}
+export class AboutComponent {}
