+++ src/app/about/presentation/contact-social-icons/contact-social-icons.component.scss
@@ -1,5 +1,5 @@
-@use 'breakpoints';
-@use 'margins';
+@use '../../../../sass/breakpoints';
+@use '../../../../sass/margins';
 
 :host {
   ul {
