+++ src/app/about/presentation/contact-social-icons/contact-social-icons.component.spec.ts
@@ -3,9 +3,9 @@ import { ComponentFixture, TestBed } from '@angular/core/testing'
 import { By } from '@angular/platform-browser'
 import { FaIconComponent } from '@fortawesome/angular-fontawesome'
 import { MockProvider } from 'ng-mocks'
-import { getComponentSelector } from '../../../test/helpers/component-testers'
-import { METADATA } from '../../common/injection-tokens'
-import { Metadata } from '../../metadata'
+import { getComponentSelector } from '../../../../test/helpers/component-testers'
+import { METADATA } from '../../../common/injection-tokens'
+import { Metadata } from '../../../metadata'
 
 import { ContactSocialIconsComponent } from './contact-social-icons.component'
