+++ src/app/about/presentation/contact-traditional-icons/contact-traditional-icons.component.scss
@@ -1,6 +1,6 @@
-@use 'breakpoints';
-@use 'margins';
-@use 'material-symbols';
+@use '../../../../sass/breakpoints';
+@use '../../../../sass/margins';
+@use '../../../../sass/material-symbols';
 
 :host {
   ul {
