+++ src/app/about/presentation/contact-traditional-icons/contact-traditional-icons.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core'
-import { Call, Email, MyLocation } from '../../material-symbols'
-import { MATERIAL_SYMBOLS_CLASS } from '../../common/material-symbols'
+import { Call, Email, MyLocation } from '../../../material-symbols'
+import { MATERIAL_SYMBOLS_CLASS } from '../../../common/material-symbols'
 
 @Component({
   selector: 'app-contact-traditional-icons',
