+++ src/app/about/presentation/presentation.component.scss
@@ -1,0 +1,52 @@
+@use 'borders';
+@use 'margins';
+@use 'paddings';
+@use 'breakpoints';
+@use 'page';
+
+:host {
+  @include page.full-size;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+
+  .contents {
+    display: flex;
+    flex-direction: column;
+    gap: margins.$l;
+    padding: paddings.$m paddings.$l;
+
+    .profile {
+      display: flex;
+      flex-direction: column;
+      flex-wrap: wrap;
+      gap: margins.$s;
+
+      h1 > span.visible-header {
+        display: flex;
+        flex-direction: column;
+
+        .real-name {
+          font-size: 1.75rem;
+          font-weight: bold;
+        }
+        code {
+          font-size: 1.5rem;
+        }
+      }
+
+      h2 {
+        font-size: 1.5rem;
+      }
+
+      .contacts {
+        display: flex;
+        gap: margins.$m;
+        flex-wrap: wrap;
+        flex-direction: row;
+        margin-top: margins.$s;
+      }
+    }
+  }
+}
