+++ src/app/about/presentation/presentation.component.ts
@@ -1,0 +1,20 @@
+import { Component, Inject } from '@angular/core'
+import { DescriptionLine, Metadata } from '../../metadata'
+import { METADATA } from '../../common/injection-tokens'
+
+@Component({
+  selector: 'app-presentation',
+  templateUrl: './presentation.component.html',
+  styleUrls: ['./presentation.component.scss'],
+})
+export class PresentationComponent {
+  public readonly realName = this.metadata.realName
+  public readonly nickname = this.metadata.nickname
+  public readonly title = this.metadata.title
+  public readonly rootLine = new DescriptionLine(
+    undefined,
+    this.metadata.descriptionLines,
+  )
+
+  constructor(@Inject(METADATA) private metadata: Metadata) {}
+}
