+++ src/app/about/presentation/profile-picture/profile-picture.component.html
@@ -5,7 +5,7 @@
   }} wearing the same and in the same pose. But appears surprised"
   class="huh"
   height="128"
-  ngSrc="assets/img/profile_huh.png"
+  ngSrc="../../../../assets/img/profile_huh.png"
   width="128"
 />
 <img
@@ -14,7 +14,7 @@
   }}. Slightly smiling. Wears 80'ish glasses and a green and black plaid shirt"
   class="main"
   height="128"
-  ngSrc="assets/img/profile.png"
+  ngSrc="../../../../assets/img/profile.png"
   priority
   width="128"
   [tabindex]="hasBeenFocused ? -1 : 0"
