+++ src/app/about/presentation/profile-picture/profile-picture.component.scss
@@ -1,6 +1,6 @@
-@use 'touch-or-pointer';
-@use 'margins';
-@use 'paddings';
+@use '../../../../sass/touch-or-pointer';
+@use '../../../../sass/margins';
+@use '../../../../sass/paddings';
 
 $picture-size: 128px;
