+++ src/app/about/presentation/profile-picture/profile-picture.component.spec.ts
@@ -1,7 +1,7 @@
 import { DebugElement } from '@angular/core'
 import { ComponentFixture, TestBed } from '@angular/core/testing'
 import { By } from '@angular/platform-browser'
-import { expectIsHidden } from '../../../test/helpers/visibility'
+import { expectIsHidden } from '../../../../test/helpers/visibility'
 
 import { ProfilePictureComponent } from './profile-picture.component'
